Fix all linter warnings (#15684)
* add updated linter config

* fix whitespace linting

* fix usestdlibvars linter

* fix lint wastedassign

* address unparam linter - container concurrency wasn't being used when assigning pods to activator

* fix tenv linter

* fix stylecheck linter

* fix staticheck linter

* fix perfsprint linter

* fix nosprintfhostport

* f perfsprint

* fix nolintlint - remove nolint directives that had no effect

* fix noctx linter

* fix gomodguard linter - drop uber atomic package

* fix nilerr linter

* fix mirror linter

* fix loggercheck linter

* fix govet linter

* fix gosimple linter

* fix gosec linter

* fix unconvert linter

* fix perfsprint linter

* fix errname linter

* fix errlint linter

* fix bodyclose linter

* fix canonicalheader linter

* fix durationcheck linter

* fix copyloopvar linter

* fix fatcontext linter

* fix gocritic-commentFormatting

* fix gocritic linter

* fix intrange linter

* fix prealloc linter

* fix flakey kpa test

* fix gofumpt linter

+++ pkg/reconciler/autoscaling/hpa/hpa.go
@@ -98,7 +98,7 @@ func (c *Reconciler) ReconcileKind(ctx context.Context, pa *autoscalingv1alpha1.
 		// as initialized until the current replicas are >= the min-scale value.
 		if !pa.Status.IsScaleTargetInitialized() {
 			ms := activeThreshold(ctx, pa)
-			if hpa.Status.CurrentReplicas >= int32(ms) {
+			if hpa.Status.CurrentReplicas >= ms {
 				pa.Status.MarkScaleTargetInitialized()
 			}
 		}
@@ -113,14 +113,14 @@ func (c *Reconciler) ReconcileKind(ctx context.Context, pa *autoscalingv1alpha1.
 }
 
 // activeThreshold returns the scale required for the pa to be marked Active
-func activeThreshold(ctx context.Context, pa *autoscalingv1alpha1.PodAutoscaler) int {
+func activeThreshold(ctx context.Context, pa *autoscalingv1alpha1.PodAutoscaler) int32 {
 	asConfig := config.FromContext(ctx).Autoscaler
 	min, _ := pa.ScaleBounds(asConfig)
 	if !pa.Status.IsScaleTargetInitialized() {
 		initialScale := getInitialScale(asConfig, pa)
-		return int(intMax(min, initialScale))
+		return intMax(min, initialScale)
 	}
-	return int(intMax(min, 1))
+	return intMax(min, 1)
 }
 
 // getInitialScale returns the calculated initial scale based on the autoscaler
